+++ PizzaBox.Testing/Specs/StoreTests.cs
@@ -32,7 +32,7 @@ namespace PizzaBox.Testing.Specs
     {
       var actual = sut.GetId(n);
 
-      Assert.IsType<long>(actual);
+      Assert.IsType<int>(actual);
     }
 
     [Theory]
@@ -41,7 +41,7 @@ namespace PizzaBox.Testing.Specs
     {
       var actual = sut.GetId(n, p);
 
-      Assert.IsType<long>(actual);
+      Assert.IsType<int>(actual);
     }
 
     [Theory]
@@ -65,7 +65,7 @@ namespace PizzaBox.Testing.Specs
     [Theory]
     [InlineData(1)]
     [InlineData(2)]
-    public void Test_RepositoryGetName(long l)
+    public void Test_RepositoryGetName(int l)
     {
       var actual = sut.GetName(l);
